fixed gcc warnings

+++ test/paste_output_test.cc
@@ -57,7 +57,7 @@ void AddRow(const std::string& qseqid, const std::string& sseqid,
      << '\t' << a.Length() << '\t' << a.Qseq() << '\t' << a.Sseq()
      << '\t' << a.Pident() << '\t' << a.RawScore() << '\t' << a.Bitscore()
      << '\t' << a.Evalue() << '\t' << a.PastedIdentifiers().at(0);
-  for (int i = 1; i < a.PastedIdentifiers().size(); ++i) {
+  for (int i = 1; i < static_cast<int>(a.PastedIdentifiers().size()); ++i) {
     ss << ',' << a.PastedIdentifiers().at(i);
   }
 }
